requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
numpy==2.3.2
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
plotly==6.0.1
shiny==1.5.0
shinywidgets==0.7.0
shinyswatch==0.9.0
joblib==1.5.2
psycopg[binary]==3.2.10
scikit-learn==1.7.1
